fix: simplify is underage and save member cover ISC handlers
They now take an UID instead of a Firebase *auth.Token.

The Firebase *auth.Token is too complex a structure to serialize via
JSON for this purpose. In addition, it cannot be serialized and
deserialized correctly / fully over JSON because it contains some
unexported fields e.g the JWT claims fields.

This limitation may lead to a downstream user of the auth token making
assumptions about it that are not correct e.g assuming that they can
read JWT claims from the deserialized *auth.Token.

As a side effect, the payloads sent when saving member covers and
checking if a member is underage have also gotten smaller.