fix : supplier edi login
- edi login should return either supplier profile or branchconnection.
when supplier profile is not nil, it means the provider that the
supplier has just logged in to has only one(1) location and the backend
has automatically set the branch as the default supplier locations.
when the branchconnection is not nil, it means the provider that the
supplier has just logged in to has more that one location hence we
return the list to location to the frontend which will prompt the
supplier to select one location from the list as their default location

- supplierSetDefaultLocation returns the updated supplier profile
instead of a simple boolean. The updated supplier profile will be used
by the frontend to update itself correctly

Signed-off-by: DavidDexter <dmwangi@savannahinformatics.com>