[ci-skip]feat: port user optout checks from the engagement service
I've added logic to check if a given user's phone number is opted out of
our marketing SMSs or not. This logic is ported from the engagement
service.

These changes ensures that the two services (engagement and profile) do
not share a database table (breaks the distributed monolith).

It includes a firestore query to check for opt out condition given a
phone number and an ISC endpoint to expose the response returned.

Signed-off-by: Dumbledore <mathenge@healthcloud.co.ke>