fix: ensure tests collections are purged after testing
Our database has for a long time been filled with leaked test
collections that are not purged when tests are tearing down.

This has been achieved by adding all test collections to our purge list
and a NOTE for developers to avoid the possibility of deleting an import
prod/testing collection during testing.

Signed-off-by: Dumbledore <mathenge@healthcloud.co.ke>