feat: implement a Stream webhook listener
This feature is primarily used to send FCM notifications on myCareHub
professional application. It works by subscribing to getStream events
via a webhook API. When an event e.g Sending a new message to a channel
is triggered, the payload will be sent to our webhook listener API. The
payload will be processed and FCMs will be sent to the staff users
within the channel that the new message was sent.