fix: search clients using phonenumbers without the country code
Phonenumbers are stored in the database with the country code i.e `+254711223344`. To ensure successful searching for Kenyan clients using the number format '0711223344',
we've made a thoughtful improvement by removing the first occurrence of '0' from the provided number.
This allows us to efficiently search for the rest of the number using PostgreSQL's pattern matching capabilities, which seamlessly handles the remaining matching process. As a result, we can now find the correct records, enabling smoother and more accurate searches for Kenyan clients in the database :-)